Implement data share by email feature. Refactor expense object and introduce currency by virtue of locale

+++ app/src/main/java/com/example/android/shissona/database/Expense.kt
@@ -12,7 +12,7 @@ data class Expense(
     val id: Int? = 0,
 
     @ColumnInfo
-    val expensePrice: Int,
+    val expensePrice: Float,
 
     @ColumnInfo
     val expenseType: Int,
